+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 * Added a new line to Redbox\Cli\Arguments::usage() so it would look better. fixed #2 
 * Added Redbox\Cli\Manager::hasDefaultValue() so you know if the default value was set #fixed #5
 * Removed the comment about default arguments and if its a good thing in Redbox\Cli\Arguments\Parser::parse()
-* Changed the comments in examples/defaultvalue.php and examples/default.php to mimic the test case better.
-
+* Changed the comments in examples/defaultvalue.php and examples/default.php to mimic the testcase better.
+* Added examples/hasdefaultvalue.php to show the Redbox\Cli\Manager::hasDefaultValue() fature.
 
 ## 1.1 - 2015-13-12